+++ gatac/tl/__init__.py
@@ -5,6 +5,7 @@ GATAC tools module - analysis functions for ATAC-seq data.
 from .peaks import call_peaks, merge_peaks, make_peak_matrix
 from .motif import (
     motif_enrichment,
+    motif_enrichment_regression,
     read_motifs,
     parse_meme,
     DNAMotif,
@@ -27,6 +28,7 @@ __all__ = [
     "merge_peaks", 
     "make_peak_matrix",
     "motif_enrichment",
+    "motif_enrichment_regression",
     "gsea_motif_enrichment",
     "sample_gc_matched_background",
     "read_motifs",